backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
certifi==2026.2.25
distlib==0.4.0
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
filelock==3.24.3
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mysql-connector-python==9.6.0
ollama==0.6.1
packaging==26.1
pillow==12.1.1
platformdirs==4.9.2
psycopg2-binary==2.9.11
pydantic==2.13.1
pydantic_core==2.46.1
PyJWT==2.12.1
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
virtualenv==20.38.0
